style: Separate 2D lights and 2D line renderers. 格式: 分离2D灯光与2D线渲染器.
