+++ src/layaAir/laya/Light2D/BaseLight2D.ts
@@ -1,22 +1,20 @@
-import { ILaya } from "../../../../ILaya";
-import { Component } from "../../../components/Component";
-import { NodeFlags } from "../../../Const";
-import { Event } from "../../../events/Event";
-import { LayaGL } from "../../../layagl/LayaGL";
-import { Color } from "../../../maths/Color";
-import { Point } from "../../../maths/Point";
-import { Rectangle } from "../../../maths/Rectangle";
-import { Vector2 } from "../../../maths/Vector2";
-import { Vector3 } from "../../../maths/Vector3";
-import { ShaderDataType } from "../../../RenderDriver/DriverDesign/RenderDevice/ShaderData";
-import { IndexFormat } from "../../../RenderEngine/RenderEnum/IndexFormat";
-import { Shader3D } from "../../../RenderEngine/RenderShader/Shader3D";
-import { BaseTexture } from "../../../resource/BaseTexture";
-import { Mesh2D, VertexMesh2D } from "../../../resource/Mesh2D";
-import { RenderTexture } from "../../../resource/RenderTexture";
-import { Utils } from "../../../utils/Utils";
-import { Scene } from "../../Scene";
-import { Sprite } from "../../Sprite";
+import { ILaya } from "../../ILaya";
+import { Component } from "../components/Component";
+import { Sprite } from "../display/Sprite";
+import { LayaGL } from "../layagl/LayaGL";
+import { Color } from "../maths/Color";
+import { Point } from "../maths/Point";
+import { Rectangle } from "../maths/Rectangle";
+import { Vector2 } from "../maths/Vector2";
+import { Vector3 } from "../maths/Vector3";
+import { ShaderDataType } from "../RenderDriver/DriverDesign/RenderDevice/ShaderData";
+import { IndexFormat } from "../RenderEngine/RenderEnum/IndexFormat";
+import { Shader3D } from "../RenderEngine/RenderShader/Shader3D";
+import { BaseTexture } from "../resource/BaseTexture";
+import { Mesh2D, VertexMesh2D } from "../resource/Mesh2D";
+import { RenderTexture } from "../resource/RenderTexture";
+import { Utils } from "../utils/Utils";
+import { Event } from "../events/Event";
 import { Light2DManager } from "./Light2DManager";
 
 /**
