+++ src/layaAir/laya/Light2D/FreeformLight2D.ts
@@ -1,24 +1,24 @@
-import { ILaya } from "../../../../ILaya";
-import { LayaGL } from "../../../layagl/LayaGL";
-import { Color } from "../../../maths/Color";
-import { Matrix } from "../../../maths/Matrix";
-import { Point } from "../../../maths/Point";
-import { Rectangle } from "../../../maths/Rectangle";
-import { Vector2 } from "../../../maths/Vector2";
-import { Vector3 } from "../../../maths/Vector3";
-import { RenderState } from "../../../RenderDriver/RenderModuleData/Design/RenderState";
-import { RenderTargetFormat } from "../../../RenderEngine/RenderEnum/RenderTargetFormat";
-import { WrapMode } from "../../../RenderEngine/RenderEnum/WrapMode";
-import { Shader3D } from "../../../RenderEngine/RenderShader/Shader3D";
-import { Material } from "../../../resource/Material";
-import { Mesh2D } from "../../../resource/Mesh2D";
-import { RenderTexture } from "../../../resource/RenderTexture";
-import { RenderTexture2D } from "../../../resource/RenderTexture2D";
-import { Texture2D } from "../../../resource/Texture2D";
-import { Sprite } from "../../Sprite";
-import { CommandBuffer2D } from "../RenderCMD2D/CommandBuffer2D";
-import { DrawMesh2DCMD } from "../RenderCMD2D/DrawMesh2DCMD";
-import { Set2DRTCMD } from "../RenderCMD2D/Set2DRenderTargetCMD";
+import { ILaya } from "../../ILaya";
+import { CommandBuffer2D } from "../display/Scene2DSpecial/RenderCMD2D/CommandBuffer2D";
+import { DrawMesh2DCMD } from "../display/Scene2DSpecial/RenderCMD2D/DrawMesh2DCMD";
+import { Set2DRTCMD } from "../display/Scene2DSpecial/RenderCMD2D/Set2DRenderTargetCMD";
+import { Sprite } from "../display/Sprite";
+import { LayaGL } from "../layagl/LayaGL";
+import { Color } from "../maths/Color";
+import { Matrix } from "../maths/Matrix";
+import { Point } from "../maths/Point";
+import { Rectangle } from "../maths/Rectangle";
+import { Vector2 } from "../maths/Vector2";
+import { Vector3 } from "../maths/Vector3";
+import { RenderState } from "../RenderDriver/RenderModuleData/Design/RenderState";
+import { RenderTargetFormat } from "../RenderEngine/RenderEnum/RenderTargetFormat";
+import { WrapMode } from "../RenderEngine/RenderEnum/WrapMode";
+import { Shader3D } from "../RenderEngine/RenderShader/Shader3D";
+import { Material } from "../resource/Material";
+import { Mesh2D } from "../resource/Mesh2D";
+import { RenderTexture } from "../resource/RenderTexture";
+import { RenderTexture2D } from "../resource/RenderTexture2D";
+import { Texture2D } from "../resource/Texture2D";
 import { BaseLight2D, Light2DType } from "./BaseLight2D";
 import { Light2DManager } from "./Light2DManager";
 import { PolygonPoint2D } from "./PolygonPoint2D";
