+++ src/layaAir/laya/Light2D/Light2DConfig.ts
@@ -1,4 +1,4 @@
-import { Color } from "../../../maths/Color";
+import { Color } from "../maths/Color";
 
 /**
  * 2D灯光全局配置参数
