+++ src/layaAir/laya/Light2D/Light2DManager.ts
@@ -1,28 +1,27 @@
-import { ILaya } from "../../../../ILaya";
-import { Laya } from "../../../../Laya";
-import { IElementComponentManager } from "../../../components/IScenceComponentManager";
-import { Event } from "../../../events/Event";
-import { LayaGL } from "../../../layagl/LayaGL";
-import { Color } from "../../../maths/Color";
-import { Matrix } from "../../../maths/Matrix";
-import { Point } from "../../../maths/Point";
-import { Rectangle } from "../../../maths/Rectangle";
-import { Vector2 } from "../../../maths/Vector2";
-import { Vector3 } from "../../../maths/Vector3";
-import { Vector4 } from "../../../maths/Vector4";
-import { BaseRenderNode2D } from "../../../NodeRender2D/BaseRenderNode2D";
-import { ShaderData, ShaderDataType } from "../../../RenderDriver/DriverDesign/RenderDevice/ShaderData";
-import { IndexFormat } from "../../../RenderEngine/RenderEnum/IndexFormat";
-import { RenderTargetFormat } from "../../../RenderEngine/RenderEnum/RenderTargetFormat";
-import { WrapMode } from "../../../RenderEngine/RenderEnum/WrapMode";
-import { Shader3D } from "../../../RenderEngine/RenderShader/Shader3D";
-import { Context } from "../../../renders/Context";
-import { Mesh2D, VertexMesh2D } from "../../../resource/Mesh2D";
-import { RenderTexture2D } from "../../../resource/RenderTexture2D";
-import { Pool } from "../../../utils/Pool";
-import { Utils } from "../../../utils/Utils";
-import { RenderState2D } from "../../../webgl/utils/RenderState2D";
-import { ILight2DManager, Scene } from "../../Scene";
+import { Context } from "vm";
+import { ILaya } from "../../ILaya";
+import { Laya } from "../../Laya";
+import { IElementComponentManager } from "../components/IScenceComponentManager";
+import { ILight2DManager, Scene } from "../display/Scene";
+import { LayaGL } from "../layagl/LayaGL";
+import { Color } from "../maths/Color";
+import { Matrix } from "../maths/Matrix";
+import { Point } from "../maths/Point";
+import { Rectangle } from "../maths/Rectangle";
+import { Vector2 } from "../maths/Vector2";
+import { Vector3 } from "../maths/Vector3";
+import { Vector4 } from "../maths/Vector4";
+import { BaseRenderNode2D } from "../NodeRender2D/BaseRenderNode2D";
+import { ShaderDataType, ShaderData } from "../RenderDriver/DriverDesign/RenderDevice/ShaderData";
+import { IndexFormat } from "../RenderEngine/RenderEnum/IndexFormat";
+import { RenderTargetFormat } from "../RenderEngine/RenderEnum/RenderTargetFormat";
+import { WrapMode } from "../RenderEngine/RenderEnum/WrapMode";
+import { Shader3D } from "../RenderEngine/RenderShader/Shader3D";
+import { Mesh2D, VertexMesh2D } from "../resource/Mesh2D";
+import { RenderTexture2D } from "../resource/RenderTexture2D";
+import { Pool } from "../utils/Pool";
+import { Utils } from "../utils/Utils";
+import { RenderState2D } from "../webgl/utils/RenderState2D";
 import { BaseLight2D, Light2DType } from "./BaseLight2D";
 import { DirectionLight2D } from "./DirectionLight2D";
 import { Light2DConfig } from "./Light2DConfig";
@@ -31,6 +30,7 @@ import { LightLine2D } from "./LightLine2D";
 import { LightOccluder2DCore } from "./LightOccluder2DCore";
 import { Occluder2DAgent } from "./Occluder2DAgent";
 import { LightAndShadow } from "./Shader/LightAndShadow";
+import { Event } from "../events/Event";
 
 /**
  * 生成2D光影图的渲染流程
