+++ src/layaAir/laya/Light2D/Light2DRenderByCmd.ts
@@ -1,17 +1,17 @@
-import { Color } from "../../../maths/Color";
-import { Matrix } from "../../../maths/Matrix";
-import { Vector2 } from "../../../maths/Vector2";
-import { RenderState } from "../../../RenderDriver/RenderModuleData/Design/RenderState";
-import { Shader3D } from "../../../RenderEngine/RenderShader/Shader3D";
-import { BaseTexture } from "../../../resource/BaseTexture";
-import { Material } from "../../../resource/Material";
-import { Mesh2D } from "../../../resource/Mesh2D";
-import { RenderTexture2D } from "../../../resource/RenderTexture2D";
-import { Scene } from "../../Scene";
-import { Command2D } from "../RenderCMD2D/Command2D";
-import { CommandBuffer2D } from "../RenderCMD2D/CommandBuffer2D";
-import { DrawMesh2DCMD } from "../RenderCMD2D/DrawMesh2DCMD";
-import { Set2DRTCMD } from "../RenderCMD2D/Set2DRenderTargetCMD";
+import { Scene } from "../display/Scene";
+import { Command2D } from "../display/Scene2DSpecial/RenderCMD2D/Command2D";
+import { CommandBuffer2D } from "../display/Scene2DSpecial/RenderCMD2D/CommandBuffer2D";
+import { DrawMesh2DCMD } from "../display/Scene2DSpecial/RenderCMD2D/DrawMesh2DCMD";
+import { Set2DRTCMD } from "../display/Scene2DSpecial/RenderCMD2D/Set2DRenderTargetCMD";
+import { Color } from "../maths/Color";
+import { Matrix } from "../maths/Matrix";
+import { Vector2 } from "../maths/Vector2";
+import { RenderState } from "../RenderDriver/RenderModuleData/Design/RenderState";
+import { Shader3D } from "../RenderEngine/RenderShader/Shader3D";
+import { BaseTexture } from "../resource/BaseTexture";
+import { Material } from "../resource/Material";
+import { Mesh2D } from "../resource/Mesh2D";
+import { RenderTexture2D } from "../resource/RenderTexture2D";
 import { BaseLight2D } from "./BaseLight2D";
 import { Light2DManager } from "./Light2DManager";
