+++ src/layaAir/laya/Light2D/LightOccluder2D.ts
@@ -1,8 +1,7 @@
-import { Component } from "../../../components/Component";
-import { NodeFlags } from "../../../Const";
-import { Event } from "../../../events/Event";
-import { Sprite } from "../../Sprite";
+import { Component } from "../components/Component";
+import { Sprite } from "../display/Sprite";
 import { Light2DManager } from "./Light2DManager";
+import { Event } from "../events/Event";
 import { LightOccluder2DCore } from "./LightOccluder2DCore";
 import { PolygonPoint2D } from "./PolygonPoint2D";
