+++ src/layaAir/laya/Light2D/Occluder2DAgent.ts
