+++ src/layaAir/laya/Line2D/ModuleDef.ts
@@ -1,0 +1,5 @@
+import { ClassUtils } from "../utils/ClassUtils";
+import { Line2DRender } from "./Line2DRender";
+
+let c = ClassUtils.regClass;
+c("Line2DRender", Line2DRender);
